Implemented update and delete functions for logged in users. Quiz came along on its own. Congrats. Version 0.07. Next up is a UX audit to start improving the overall experience.

+++ src/components/DeleteCard.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { firestore } from '../firebase'
 
-const DeleteCard = ({id}) => {
+const DeleteCard = ({id, user}) => {
 
     const remove = async (e) => {
         e.preventDefault();
-        await firestore.doc(`cards/${id}`).delete(); 
+        await firestore.doc(`users/${user.uid}/cards/${id}`).delete();
     }
 
     return(
